feat(federation): vault replication pipeline with re-encryption and signature binding
core-crypto:
- replication_envelope_aad() shared AAD construction for sender and
  receiver. Layout: batch_hash || timestamp_secs (LE u64) ||
  session_id_len (LE u32) || session_id. Length prefix prevents
  canonicalization collision between distinct session IDs.
- REPLICATION_HKDF_CONTEXT = b"opensesame:vault:replication:v1" exported
  as wire-format constant. Imported by both daemon-network sender and
  daemon-secrets receiver — compile-time agreement.
- tests/replication_envelope.rs: 9 tests covering seal/open round-trip,
  wrong-key rejection, tampered ciphertext/timestamp/session_id/
  batch_hash rejection, empty session_id, 64KB payload, AAD
  determinism, length-prefix separation.

core-types:
- HlcTimestamp wall_secs and counter promoted from u32 to u64.
  HlcTimestamp::tick and HlcTimestamp::receive signatures updated.
  u32::MAX overflow check replaced with u64::MAX.
- VaultLogEntryReceived: profile_id field renamed to
  peer_installation_id (carries relay peer installation UUID, not
  a profile ID).
- VaultReplicationPullRequest: profile_id (Uuid) replaced with
  profile_name (String); peer_id field added (echoed in response for
  caller-side watermark caching keyed by installation ID).
- VaultReplicationPullResponse: profile_id replaced with profile_name;
  peer_id field added; last_hlc_json field added (HLC of last returned
  entry as JSON, used by daemon-network to set since_watermark on next
  pull).
- ReplicationPullProgressUpdate event variant added: { peer_id,
  profile_name, last_hlc_json }. Published by daemon-network to
  daemon-secrets after each pull response.

core-config:
- VaultSyncConfig.compaction_retention_secs (u64) added with
  default 604_800 (7 days).
- VaultSyncConfig added as vault_sync field on Config struct.

daemon-secrets crate split:
- src/lib.rs added; all modules declared pub for integration test
  access through the library boundary.
- src/main.rs reduced to binary-only logic; uses daemon_secrets::*.

daemon-secrets/src/crud.rs:
- SIGNING_SEED and NETWORK_PRIVATE_KEY: RwLock<Option<Zeroizing<[u8;
  32]>>>. Closure-based accessors with_signing_seed/with_network_private_key
  keep secret material inside the lock scope.
- set_vault_log validates installation UUID before setting INSTALL_ID
  OnceLock. Returns Err on validation failure — OnceLock never poisoned
  with invalid value.
- Reactive signing seed cache in handle_secret_set for _signing-seed
  and _network-identity-private writes.

daemon-secrets/src/dispatch.rs:
- process_received_batch: signature verification FIRST via
  validate_entry_structure before any field reads. Differentiates
  RejectedEntry (warn audit) from InvalidSignature (debug drop).
- value_hash binding: BLAKE3(received_value) verified against signed
  value_hash. Empty value_hash on set operations rejected. Hash
  mismatch defers with exponential backoff, no HWM advance.
- build_pull_response: verifies current vault value matches signed
  value_hash before attaching. Mismatched values omitted.
- ReplicationPullProgressUpdate handler gated on
  verified_sender_name == "daemon-network".
- apply_deferred_entries: deferred sets left unapplied (no value in DB),
  deferred deletes applied on profile unlock.

daemon-secrets/src/vault_log.rs:
- Schema versioning via PRAGMA user_version. Mismatched versions refuse
  to start with clear error — no silent data destruction.
- Split tables: pull_progress (source_peer, profile) for sync progress,
  entry_replay_hwm (author, profile) for replay protection.
- canonical_sign_payload gains value_hash field. write_local_entry
  computes BLAKE3 of value_bytes for all set operations.
- insert_received_entry normalizes operation_json to inner object.
- validate_entry_structure: signature before content checks. System-key
  rejection uses RejectedEntry variant.
- HLC fields stored as i64 in SQL, u64 in Rust structs.

daemon-network:
- Per-destination re-encryption: ephemeral X25519 ECDH per session,
  HKDF-BLAKE2b key derivation, ChaCha20-Poly1305 seal with
  replication_envelope_aad AAD binding.
- Inbound replication routing in UDP/TCP Data handlers: 256KB frame cap,
  UTF-8 validation, TOFU pin check (reject Unpinned/Revoked),
  per-installation rate limiting (10/s burst 50).
- run_replication_pull: 60s periodic pull with cached per-peer watermarks.
- Startup validation: refuse if default_profile empty or config load fails.
  Validate peer installation UUID is non-nil before IPC forwarding.

daemon-secrets/tests/replication_pipeline.rs: 28 integration tests
covering positive and negative paths through the replication pipeline.

+++ core-types/src/network.rs
@@ -161,10 +161,10 @@ pub enum VaultLogOp {
 /// are declared in sort-priority order.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, Serialize, Deserialize)]
 pub struct HlcTimestamp {
-    /// Coarse wall clock (Unix epoch seconds). `u32` covers until 2106.
-    pub wall_secs: u32,
+    /// Coarse wall clock (Unix epoch seconds). `u64` avoids Y2106 overflow.
+    pub wall_secs: u64,
     /// Per-wall-second logical counter for ordering within the same second.
-    pub counter: u32,
+    pub counter: u64,
     /// First 8 bytes of `InstallationId.id` UUID. Tiebreaker for identical
     /// `(wall_secs, counter)` pairs from different nodes.
     pub node_id: [u8; 8],
@@ -185,13 +185,13 @@ impl HlcTimestamp {
     ///
     /// Guarantees the returned timestamp is strictly greater than the
     /// previous value of `local`.
-    pub fn tick(local: &mut Self, wall_now: u32) -> Self {
+    pub fn tick(local: &mut Self, wall_now: u64) -> Self {
         if wall_now > local.wall_secs {
             local.wall_secs = wall_now;
             local.counter = 0;
         } else {
             local.counter = local.counter.saturating_add(1);
-            if local.counter == u32::MAX {
+            if local.counter == u64::MAX {
                 local.wall_secs = local.wall_secs.saturating_add(1);
                 local.counter = 0;
             }
@@ -203,7 +203,7 @@ impl HlcTimestamp {
     ///
     /// Guarantees the returned timestamp is strictly greater than both
     /// the previous `local` value and `recv_ts`.
-    pub fn receive(local: &mut Self, recv_ts: &Self, wall_now: u32) -> Self {
+    pub fn receive(local: &mut Self, recv_ts: &Self, wall_now: u64) -> Self {
         let max_wall = wall_now.max(local.wall_secs).max(recv_ts.wall_secs);
         if max_wall > local.wall_secs && max_wall > recv_ts.wall_secs {
             local.wall_secs = max_wall;
